+++ prisma/schema.prisma
@@ -12,12 +12,13 @@ datasource db {
 }
 
 model Post {
-  id        String   @id @default(cuid())
-  createdAt DateTime @default(now())
-  content   String   @db.VarChar(255)
-  authorId  String
+  id         String   @id @default(cuid())
+  createdAt  DateTime @default(now())
+  content    String   @db.VarChar(255)
+  authorId   String
+  replyingTo String   @default("")
 
-  @@index([authorId])
+  @@index([authorId, replyingTo])
 }
 
 model Like {
