+++ src/pages/post/[id].tsx
@@ -5,11 +5,15 @@ import { api } from "~/utils/api";
 import { PageLayout } from "~/components/layout";
 import { generateSSGHelper } from "~/server/helpers/ssgHelper";
 import { PostView } from "~/components/postview";
-
+import { CreatePostWizard } from "~/components/create-post-wizard";
 
 const SinglePostPage: NextPage<{ id: string }> = ({ id }) => {
   const { data } = api.posts.getById.useQuery({
     id
+  });
+
+  const { data: replies } = api.posts.getAllReplies.useQuery({
+    postId: id
   });
 
   if (!data) return <div>404</div>;
@@ -21,6 +25,16 @@ const SinglePostPage: NextPage<{ id: string }> = ({ id }) => {
       </Head>
       <PageLayout>
         <PostView {...data} />
+        <div className="p-8">
+          <CreatePostWizard replyingTo={data.post.id} />
+        </div>
+        <div className="border-b border-slate-400">
+          {replies?.map(reply => (
+            <div key={reply.post.id}>
+              <PostView {...reply} />
+            </div>
+          ))}
+        </div>
       </PageLayout>
     </>
   );
@@ -34,6 +48,7 @@ export const getStaticProps: GetStaticProps = async (context) => {
   if (typeof id !== "string") throw new Error("No id");
 
   await ssg.posts.getById.prefetch({ id })
+  await ssg.posts.getAllReplies.prefetch({ postId: id })
 
   return {
     props: {
